Moved OpenGL logger to a separate file
